backend/requirements.txt
fastapi
uvicorn
python-multipart
transformers
sentence-transformers
torch
qdrant-client
pdfplumber
python-docx
wordfreq
numpy
pandas
tqdm
python-dotenv
pydantic
pydantic-settings
loguru
requests
groq
pytest
langchain==1.0.3
langgraph
langchain-core==0.3.79 
langchain-google-genai==2.0.2
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
sqlalchemy
psycopg2-binary
pymysql